Rename ZUIAvatar to ZUIPersonAvatar.
